+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# dev
+
+# 1.1.0
+standardization: handle malformed laz input ("Global encoding WKT flag not set for point format 6 - 10")
+color: extract unlock module from colorization and rename colorization function
+
 # 1.0.0
 - first public version
 - docker: Use staged build to reduce docker image size